[Fix] resolution when `filename` option is passed
PR #162 introduced a `parent` variable that contains either the full
filename of the requiring module or the `basedir`. Some of the places
where `basedir` was used were updated to use `parent` instead, but
that's not correct; the `path.resolve()` calls would receive the full
filename instead of the directory name.

This patch reverts the improper `parent` uses to `basedir`.