[New] resolve.sync can resolve to options.fallback if resolution fails
Added a new optional field `fallback` in `resolve.sync`'s `options`
parameter. If given,  `fallback` must be a `string` value for it to be
considered in the resolution process. It can be an empty string.

If provided, and if the module resolution is about to fail, the fallback
value is returned instead, and no exception is thrown.

Old users of the API won't be affected by this change, and new user will
be able to avoid a possibly costly exception throw -- especially relevant
in scenarios where there's a considerable amount of modules
(direct or transitive) to resolve.

Please see `tests/fallback.js` for a brief example.